Add files via upload
A REXX environment on a Windows should be in place to run this tool. The open source: OOREXX (http://www.oorexx.org/download.html) is the best choice.

Also a MicroSoft Excel is needed. When the office version is upgraded, some changes must be made in the following REXX Exec: “IOCDS 2 XLS.REX”